requirements.txt
google-generativeai
python-dotenv
streamlit
SpeechRecognition
pyttsx3
pyaudio
wikipedia